move utils to separate module
